% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{calculate_angle}
\alias{calculate_angle}
\title{Calculate Angle Between Two Vectors}
\usage{
calculate_angle(u, v)
}
\arguments{
\item{u}{Numeric vector representing the first vector.}

\item{v}{Numeric vector representing the second vector.}
}
\value{
Numeric value representing the angle in degrees.
}
\description{
Calculate Angle Between Two Vectors
}
