% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{calculate_angles_for_track}
\alias{calculate_angles_for_track}
\title{Calculate Angles for a Track}
\usage{
calculate_angles_for_track(track_data, x, y)
}
\arguments{
\item{track_data}{Data frame containing the track data with columns 'x' and 'y'}

\item{x}{The colum name with the x-coordinates of the 2D motion capture}

\item{y}{The colum name with the y-coordinates of the 2D motion capture}
}
\value{
Numeric vector of angles for the track.
}
\description{
Calculate Angles for a Track
}
