% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{complement_angle}
\alias{complement_angle}
\title{Transform angles to their complement angle}
\usage{
complement_angle(x, ref = 180)
}
\arguments{
\item{x}{Numeric. Angles to transform into their complement respective to \code{ref}}

\item{ref}{Numeric. Number from which to take the complement.}
}
\value{
Numeric vector of complements.
}
\description{
Transform angles to their complement angle
}
