% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gait.R
\docType{data}
\name{gait}
\alias{gait}
\title{Basic 2D gait analysis dataset}
\format{
A data.frame, data.table with 168 rows and 5 columns:
\describe{
\item{nr}{The row number}
\item{joint}{The id for the joint}
\item{frame}{the frame number for the respective joint}
\item{x_coord}{The X coordinates for the position of the joint in the specific frame}
\item{y_coord}{The Y coordinates for the position of the joint in the specific frame}
}
}
\usage{
gait
}
\description{
This is an example dataset from a motion capture of a gait, with frames por the
position of each major joint of the right lower limb
}
\keyword{datasets}
