% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-motion.R
\name{plot_motion}
\alias{plot_motion}
\title{Visualize the motion capture across frames}
\usage{
plot_motion(data, x, y, frames)
}
\arguments{
\item{data}{Dataset with motion capture data.}

\item{x}{Unquoted variable name of the x coordinate variable.}

\item{y}{Unquoted variable name of the y coordinate variable.}

\item{frames}{Unquoted variable name with the frame ids for each joint.}
}
\value{
ggplot2 object
}
\description{
Visualize the motion capture across frames
}
