% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{to_degrees}
\alias{to_degrees}
\title{Transform radians to degrees}
\usage{
to_degrees(x)
}
\arguments{
\item{x}{Numeric. Radians to transform to degrees.}
}
\value{
Numeric vector.
}
\description{
Transform radians to degrees
}
