% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{to_radians}
\alias{to_radians}
\title{Transform degrees to radians}
\usage{
to_radians(x)
}
\arguments{
\item{x}{Numeric. Degrees to transform to radians.}
}
\value{
Numeric vector.
}
\description{
Transform degrees to radians
}
