% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-motion.R
\name{plot_degrees}
\alias{plot_degrees}
\title{Visualize the angles between joints across frames}
\usage{
plot_degrees(data, x, y, joint_ids, frames, plot = TRUE)
}
\arguments{
\item{data}{Dataset with motion capture data.}

\item{x}{Unquoted variable name of the x coordinate variable.}

\item{y}{Unquoted variable name of the y coordinate variable.}

\item{joint_ids}{Unquoted variable name with the joint ids.}

\item{frames}{Unquoted variable name with the frame ids for each joint.}

\item{plot}{Logical. Wheather to obtain the plot (TRUE) or the data with angle measurements (F).}
}
\value{
ggplot2 object
}
\description{
Visualize the angles between joints across frames
}
